items: add common function for taking damage (#888)
Resolves #876.

+++ src/game/gun/gun_misc.h
@@ -4,6 +4,7 @@
 
 #include "global/types.h"
 
+#include <stdbool.h>
 #include <stdint.h>
 
 extern WEAPON_INFO g_Weapons[NUM_WEAPONS];
@@ -14,6 +15,6 @@ void Gun_FindTargetPoint(ITEM_INFO *item, GAME_VECTOR *target);
 void Gun_AimWeapon(WEAPON_INFO *winfo, LARA_ARM *arm);
 int32_t Gun_FireWeapon(
     int32_t weapon_type, ITEM_INFO *target, ITEM_INFO *src, PHD_ANGLE *angles);
-void Gun_HitTarget(ITEM_INFO *item, GAME_VECTOR *hitpos, int32_t damage);
+void Gun_HitTarget(ITEM_INFO *item, GAME_VECTOR *hitpos, int16_t damage);
 
 void Gun_DrawFlash(LARA_GUN_TYPE weapon_type, int32_t clip);
